refactor(agent): extract the run executor from AgentRuntime
Driving a run from its first round to a settled outcome moves into
AgentRunExecutor: the catch order ADR-084 makes a hard guarantee,
ADR-103's cancellation probe, ADR-104's lease heartbeat and lease-lost
stop, ADR-111's write fence and fail-closed audit.

It no longer knows where a run came from. A request, a claimed queue row
and a resumed suspension all reach it as a handle, a trace and a closure
producing the next tool-loop result, so the ordering guarantees hold
identically on all three paths — and can be exercised without a queue
row, a resume claim or a provider in the way.

approve() and submitInput() previously each rebuilt the same sequence:
trace over the claimed handle, tool context from the run OWNER rather
than the approver or submitter (ADR-083), then the ladder around the
loop call. They now share one entry point that fixes that order in one
place and takes only the tool-loop call that differs between them.

Also repairs two things the previous extraction broke and nothing
flagged: QueuedRunFailureRecovery carried no CoversClass anywhere, so
its ~160 executed lines had dropped out of attributed coverage (both
PHPUnit configs set requireCoverageMetadata="false"), and ADR-112 still
named AgentRuntime::recoverQueuedFailure(). ADR-103's reference to the
runtime's trace hook now names the executor.
