+++ Classes/Service/Agent/AgentRuntime.php
@@ -12,7 +12,6 @@ namespace Netresearch\NrLlm\Service\Agent;
 use Closure;
 use Netresearch\NrLlm\Domain\Enum\AgentRunOutcome;
 use Netresearch\NrLlm\Domain\Enum\AgentRunStatus;
-use Netresearch\NrLlm\Domain\Enum\AgentRunTerminationReason;
 use Netresearch\NrLlm\Domain\Enum\ServiceAccountScope;
 use Netresearch\NrLlm\Domain\Enum\ToolEffect;
 use Netresearch\NrLlm\Domain\Repository\LlmConfigurationRepository;
@@ -21,28 +20,20 @@ use Netresearch\NrLlm\Domain\ValueObject\AiActorContext;
 use Netresearch\NrLlm\Domain\ValueObject\RunStep;
 use Netresearch\NrLlm\Domain\ValueObject\SuspendedRunState;
 use Netresearch\NrLlm\Domain\ValueObject\ToolLoopResult;
-use Netresearch\NrLlm\Exception\GuardrailApprovalRequiredException;
-use Netresearch\NrLlm\Exception\GuardrailViolationException;
-use Netresearch\NrLlm\Service\Agent\Exception\AuditPersistenceFailedException;
 use Netresearch\NrLlm\Service\Agent\Exception\CorruptSuspendedStateException;
 use Netresearch\NrLlm\Service\Agent\Exception\InvalidInputSubmissionException;
 use Netresearch\NrLlm\Service\Agent\Exception\RunAccessDeniedException;
 use Netresearch\NrLlm\Service\Agent\Exception\RunAlreadyResumingException;
-use Netresearch\NrLlm\Service\Agent\Exception\RunCancellationRequestedException;
 use Netresearch\NrLlm\Service\Agent\Exception\RunConfigurationGoneException;
 use Netresearch\NrLlm\Service\Agent\Exception\RunEnqueueFailedException;
-use Netresearch\NrLlm\Service\Agent\Exception\RunLeaseLostException;
 use Netresearch\NrLlm\Service\Agent\Exception\RunNotAwaitingApprovalException;
 use Netresearch\NrLlm\Service\Agent\Exception\RunNotAwaitingInputException;
 use Netresearch\NrLlm\Service\Agent\Exception\RunStateUnavailableException;
 use Netresearch\NrLlm\Service\Agent\Queue\AgentRunQueuedMessage;
 use Netresearch\NrLlm\Service\Schema\JsonSchemaValidator;
-use Netresearch\NrLlm\Service\Tool\ActingBackendUserResolver;
 use Netresearch\NrLlm\Service\Tool\ActingBackendUserResolverInterface;
 use Netresearch\NrLlm\Service\Tool\AgentRunHandle;
 use Netresearch\NrLlm\Service\Tool\AgentRunPersister;
-use Netresearch\NrLlm\Service\Tool\Exception\ToolApprovalRequiredException;
-use Netresearch\NrLlm\Service\Tool\Exception\ToolInputRequiredException;
 use Netresearch\NrLlm\Service\Tool\InputSchema;
 use Netresearch\NrLlm\Service\Tool\RunTrace;
 use Netresearch\NrLlm\Service\Tool\ToolEffectResolver;
@@ -139,7 +130,28 @@ final readonly class AgentRuntime implements AgentRuntimeInterface
         // builds one over the persister, logger and bus this runtime holds,
         // which is exactly what it was constructed with before.
         private ?QueuedRunFailureRecovery $failureRecovery = null,
+        // Drives a run from its first round to a settled outcome (ADR-084's
+        // catch order, ADR-103's cancellation probe, ADR-104's lease, ADR-111's
+        // write fence). Optional for the same reason as the collaborators
+        // above; a null builds one over the tool loop, persister, logger and
+        // the two resolvers this runtime already holds.
+        private ?AgentRunExecutor $runExecutor = null,
     ) {}
+
+    /**
+     * The executor for a run's lifecycle, falling back to one built from this
+     * runtime's own collaborators when none was injected.
+     */
+    private function runExecutor(): AgentRunExecutor
+    {
+        return $this->runExecutor ?? new AgentRunExecutor(
+            $this->toolLoop,
+            $this->persister,
+            $this->logger,
+            $this->actingBackendUserResolver,
+            $this->toolEffectResolver,
+        );
+    }
 
     /**
      * The recovery for a failed queued run, falling back to one built from this
@@ -159,24 +171,11 @@ final readonly class AgentRuntime implements AgentRuntimeInterface
         return $this->requestCodec ?? new AgentRunRequestCodec($this->configurationRepository);
     }
 
-    /**
-     * Build the tool-execution context for a run from its explicit actor: the
-     * one place the actor becomes a live acting backend user, identically on the
-     * synchronous and worker paths (ADR-083), so no tool reads the ambient
-     * `$GLOBALS['BE_USER']`.
-     */
-    private function toolContext(AiActorContext $actor): ToolExecutionContext
-    {
-        $resolver = $this->actingBackendUserResolver ?? new ActingBackendUserResolver();
-
-        return new ToolExecutionContext($actor, $resolver->resolve($actor));
-    }
-
     public function run(AgentRunRequest $request, ?Closure $onStep = null): AgentRunResult
     {
         $handle = $this->persister->begin($request->configuration, $request->actor->backendUserUid);
 
-        return $this->executeRequest($request, $handle, $onStep);
+        return $this->runExecutor()->executeRequest($request, $handle, $onStep);
     }
 
     public function enqueue(AgentRunRequest $request): string
@@ -285,54 +284,7 @@ final readonly class AgentRuntime implements AgentRuntimeInterface
             return $this->failureRecovery()->recover($handle, $runUuid, $workerIdentity, $e, $steps);
         };
 
-        return $this->executeRequest($request, $handle, $onStep, $workerIdentity, $recover);
-    }
-
-    /**
-     * The shared execution path behind {@see run()} and {@see runQueued()}:
-     * clamp the round cap, build the trace, drive the ladder.
-     *
-     * A queued run passes its worker identity as $leaseOwner (so each step
-     * boundary renews the lease and detects a reaper reclaim, ADR-104) and a
-     * $recover closure that decides retry-vs-dead-letter for a failure. An
-     * interactive run() passes neither: it holds no lease and surfaces failures
-     * to its caller unchanged.
-     *
-     * @param (Closure(RunStep): void)|null                             $onStep
-     * @param (Closure(Throwable, list<RunStep>): ?AgentRunResult)|null $recover
-     */
-    private function executeRequest(
-        AgentRunRequest $request,
-        ?AgentRunHandle $handle,
-        ?Closure $onStep,
-        ?string $leaseOwner = null,
-        ?Closure $recover = null,
-    ): AgentRunResult {
-        $maxIterations = $request->maxIterations !== null
-            ? min($request->maxIterations, self::MAX_ITERATIONS)
-            : null;
-
-        $trace = $this->trace($handle, $onStep, $request->captureRaw, $leaseOwner);
-        // Resolve the run's explicit actor to a live acting backend user ONCE,
-        // identically whether this runs synchronously or in a worker (ADR-083).
-        $context = $this->toolContext($request->actor);
-
-        return $this->execute(
-            $handle,
-            $trace,
-            fn(): ToolLoopResult => $this->toolLoop->runLoop(
-                $request->messages,
-                $request->configuration,
-                $context,
-                $request->allowedToolNames,
-                $request->options,
-                $maxIterations,
-                $trace,
-                $request->augmentation,
-            ),
-            $recover,
-            $leaseOwner,
-        );
+        return $this->runExecutor()->executeRequest($request, $handle, $onStep, $workerIdentity, $recover);
     }
 
     /**
@@ -402,18 +354,16 @@ final readonly class AgentRuntime implements AgentRuntimeInterface
         // who approved or denied, before the continuation's own events.
         $this->persister->recordApproval($handle, $decision->approved, $decision->decidedByBeUser);
 
-        $trace = $this->trace($handle, $onStep, false);
         // Tools resume under the RUN OWNER's identity (ADR-083), not whoever is
         // approving: the owner is who the queued work acts for. Reconstructed
         // from the run row's initiating uid — the same fallback the request codec
         // uses — and resolved to a live user, so authorization is identical to
         // the original synchronous/worker execution.
-        $context = $this->toolContext(AiActorContext::backendUser($run->beUser));
-
-        return $this->execute(
+        return $this->runExecutor()->executeResume(
             $handle,
-            $trace,
-            fn(): ToolLoopResult => $this->toolLoop->resume(
+            $onStep,
+            AiActorContext::backendUser($run->beUser),
+            fn(ToolExecutionContext $context, RunTrace $trace): ToolLoopResult => $this->toolLoop->resume(
                 $state,
                 $decision->approved,
                 $configuration,
@@ -488,15 +438,13 @@ final readonly class AgentRuntime implements AgentRuntimeInterface
         // submitted, before the continuation's own events — never the values.
         $this->persister->recordInput($handle, $submission->submittedByBeUser);
 
-        $trace = $this->trace($handle, $onStep, false);
         // Tools resume under the RUN OWNER's identity (ADR-083), not the
         // submitter — same rule as approve().
-        $context = $this->toolContext(AiActorContext::backendUser($run->beUser));
-
-        return $this->execute(
+        return $this->runExecutor()->executeResume(
             $handle,
-            $trace,
-            fn(): ToolLoopResult => $this->toolLoop->resumeWithInput(
+            $onStep,
+            AiActorContext::backendUser($run->beUser),
+            fn(ToolExecutionContext $context, RunTrace $trace): ToolLoopResult => $this->toolLoop->resumeWithInput(
                 $state,
                 $submission->data,
                 $configuration,
@@ -544,110 +492,6 @@ final readonly class AgentRuntime implements AgentRuntimeInterface
     }
 
     /**
-     * The trace every segment runs under: each recorded step reaches the live
-     * observer FIRST (preserving the streaming path's emit-before-persist
-     * order — a step is shown even when its persist fails), then the persisted
-     * event stream — and then the cancellation probe runs (ADR-103).
-     *
-     * The probe is what makes {@see cancel()} cooperative: the loop itself
-     * stays persistence-unaware (ADR-081), but every step boundary — after a
-     * provider response, after each tool execution, before the next round —
-     * records a step, and the probe checks the run row there. A run cancelled
-     * mid-flight therefore stops before the NEXT provider call or tool
-     * execution instead of running to completion with its outcome discarded.
-     * The step that just happened is still emitted and persisted: the audit
-     * stream stays complete up to the abort point. One indexed row read per
-     * step; steps are provider-call-slow, so the cost is negligible.
-     *
-     * For a queued run $leaseOwner is this worker's identity (ADR-104): after
-     * the cancellation check, the lease is renewed under an ownership guard
-     * BEFORE the step is persisted. If the renewal affects no row the worker no
-     * longer owns the run — the reaper reclaimed it and another worker may hold
-     * it now — so it stops WITHOUT writing the step, which would otherwise
-     * collide with the new owner's stream. Interactive runs pass null and never
-     * renew (they hold no lease).
-     *
-     * @param (Closure(RunStep): void)|null $onStep
-     */
-    private function trace(?AgentRunHandle $handle, ?Closure $onStep, bool $captureRaw, ?string $leaseOwner = null): RunTrace
-    {
-        if ($handle === null && $onStep === null) {
-            return new RunTrace(captureRaw: $captureRaw);
-        }
-
-        return new RunTrace(
-            captureRaw: $captureRaw,
-            onRecord: function (RunStep $step) use ($handle, $onStep, $leaseOwner): void {
-                // The live observer sees the step FIRST (emit-before-persist),
-                // so a step is shown even when its persistence or the checks
-                // below abort the loop.
-                if ($onStep !== null) {
-                    $onStep($step);
-                }
-                if ($handle === null) {
-                    return;
-                }
-
-                // A single indexed read drives the cancellation probe (ADR-103).
-                // findRun is fail-soft (null on a store hiccup), so a read
-                // failure can never fabricate a cancellation.
-                if ($this->persister->findRun($handle->uuid)?->statusEnum() === AgentRunStatus::CANCELLED) {
-                    // Persist the step that already happened, then stop: the
-                    // audit stream stays complete up to the abort point. A
-                    // writing tool whose audit cannot be stored fails the run
-                    // (ADR-111) even mid-cancel — an unrecorded mutation is the
-                    // more serious condition than the cancellation.
-                    $this->recordStepFailClosedForWrites($handle, $step);
-
-                    throw RunCancellationRequestedException::forRun($handle->uuid);
-                }
-
-                // Heartbeat + lease-lost guard for a worker run (ADR-104). The
-                // renewal's ownership guard is the atomic check: 0 rows means
-                // the run was reclaimed/re-claimed/terminated — stop WITHOUT
-                // recording the step, so a zombie worker cannot append an event
-                // whose sequence collides with the new owner's stream. A completed
-                // WRITE step also CLEARS the in-flight fence set before it ran
-                // (ADR-111) in the same guarded write.
-                if ($leaseOwner !== null && !$this->renewOrClearFence($handle, $leaseOwner, $step)) {
-                    throw RunLeaseLostException::forRun($handle->uuid);
-                }
-
-                $this->recordStepFailClosedForWrites($handle, $step);
-            },
-            onBeforeTool: $leaseOwner === null || $handle === null ? null : function (string $toolName) use ($handle, $leaseOwner): void {
-                // Fence a WRITE before it runs (ADR-111): stamp its effect and
-                // renew the lease so a reap mid non-idempotent-write dead-letters
-                // instead of retrying. Read-only tools need no fence — repeating
-                // them is always safe. A lost lease stops the worker before the
-                // side effect, exactly like the heartbeat guard.
-                $effect = $this->toolEffectResolver?->effectFor($toolName) ?? ToolEffect::READ_ONLY;
-                if ($effect->isWrite()
-                    && !$this->persister->markPendingEffect($handle, $leaseOwner, $effect->value, time() + self::LEASE_SECONDS)
-                ) {
-                    throw RunLeaseLostException::forRun($handle->uuid);
-                }
-            },
-        );
-    }
-
-    /**
-     * Renew the lease at a step boundary and, for a completed WRITE step, clear
-     * the in-flight fence its {@see RunTrace::beforeToolExecution()} set — both in
-     * one ownership-guarded write (ADR-111). Returns false when the worker has
-     * lost the run, so the caller stops before recording the step.
-     */
-    private function renewOrClearFence(AgentRunHandle $handle, string $leaseOwner, RunStep $step): bool
-    {
-        $leaseExpires = time() + self::LEASE_SECONDS;
-        if ($this->stepEffect($step)->isWrite()) {
-            return $this->persister->markPendingEffect($handle, $leaseOwner, '', $leaseExpires);
-        }
-
-        return $this->persister->renewLease($handle, $leaseOwner, $leaseExpires);
-    }
-
-    /**
      * Whether a run may be retried given the effect fence in flight when it
      * failed (ADR-111). Empty (no tool fenced) or an unrecognised value is safe;
      * only a fenced NON_IDEMPOTENT_WRITE — a side effect that must not repeat —
@@ -658,264 +502,4 @@ final readonly class AgentRuntime implements AgentRuntimeInterface
         return ToolEffect::tryFrom($pendingEffect)?->isSafeToRetry() ?? true;
     }
 
-    /**
-     * Persist a step, failing the run when a WRITING tool's audit event could
-     * not be stored (ADR-111). Read-only and non-tool steps stay fail-soft: a
-     * store hiccup is logged inside {@see AgentRunPersister::recordStep()} and
-     * the run continues. A write is different — an unrecorded mutation must not
-     * be waved through, so it throws {@see AuditPersistenceFailedException}
-     * (non-retryable: the write already ran once).
-     */
-    private function recordStepFailClosedForWrites(AgentRunHandle $handle, RunStep $step): void
-    {
-        $persisted = $this->persister->recordStep($handle, $step);
-        if (!$persisted && $this->stepEffect($step)->isWrite()) {
-            throw AuditPersistenceFailedException::forRun($handle->uuid, $step->toolName ?? '');
-        }
-    }
-
-    /**
-     * The side effect of the tool a step recorded (ADR-111). Only KIND_TOOL
-     * steps carry an effect; everything else is read-only. A tool name that no
-     * longer resolves is treated as the strictest effect (fail-closed), and a
-     * runtime built without the resolver (bare positional test wiring only)
-     * treats every tool as read-only.
-     */
-    private function stepEffect(RunStep $step): ToolEffect
-    {
-        if ($step->kind !== RunStep::KIND_TOOL || $this->toolEffectResolver === null) {
-            return ToolEffect::READ_ONLY;
-        }
-
-        return $this->toolEffectResolver->effectFor($step->toolName ?? '');
-    }
-
-    /**
-     * The single lifecycle ladder (ADR-101; previously copied three times in
-     * the playground controller).
-     *
-     * Every branch both settles the persisted row and marks the run settled so
-     * the finally-guard cannot touch it — including a SUCCESSFUL suspension:
-     * WAITING_FOR_APPROVAL is non-terminal, so an unguarded finally-settle
-     * would flip a resumable run to FAILED and destroy its suspended state.
-     * The finally-guard exists for the abandoned-run case (a live-stream
-     * observer dying mid-run, mirroring StreamingDispatcher) and costs the
-     * settled paths nothing.
-     *
-     * A queued run passes a $recover closure (ADR-104): in the generic failure
-     * arm, before the default settleFailed, it decides retry-vs-dead-letter and,
-     * when it handles the failure, returns the outcome that replaces FAILED.
-     *
-     * @param Closure(): ToolLoopResult                                 $loopCall
-     * @param (Closure(Throwable, list<RunStep>): ?AgentRunResult)|null $recover
-     */
-    private function execute(?AgentRunHandle $handle, RunTrace $trace, Closure $loopCall, ?Closure $recover = null, ?string $ownerGuard = null): AgentRunResult
-    {
-        $runUuid = $handle !== null ? $handle->uuid : '';
-        $settled = false;
-
-        try {
-            $result = $loopCall();
-
-            if ($handle !== null) {
-                $settledOk = $this->persister->settleCompleted($handle, $result, $ownerGuard);
-                // Ownership-guarded on the queued path: if the run was reclaimed
-                // to another worker mid-loop, this completion must not overwrite
-                // its state — stop as LEASE_LOST rather than reporting COMPLETED.
-                if ($ownerGuard !== null && !$settledOk) {
-                    $settled = true;
-
-                    return new AgentRunResult(outcome: AgentRunOutcome::LEASE_LOST, runUuid: $runUuid, steps: $trace->getSteps());
-                }
-            }
-            $settled = true;
-
-            return new AgentRunResult(
-                outcome: AgentRunOutcome::COMPLETED,
-                runUuid: $runUuid,
-                steps: $trace->getSteps(),
-                loopResult: $result,
-            );
-        } catch (RunCancellationRequestedException $cancelled) {
-            // ADR-103: the operator's cancel already won the guarded terminal
-            // transition — the row IS CANCELLED and its late settle would be
-            // discarded anyway, so none is attempted. Control flow, not
-            // failure: the loop stopped cooperatively at a step boundary.
-            $settled = true;
-            $this->logger?->info('Agent run stopped cooperatively after being cancelled', ['run' => $runUuid]);
-
-            return new AgentRunResult(
-                outcome: AgentRunOutcome::CANCELLED,
-                runUuid: $runUuid,
-                steps: $trace->getSteps(),
-                error: $cancelled,
-            );
-        } catch (ToolApprovalRequiredException $approval) {
-            // ADR-084: a called tool requires human approval — control flow,
-            // not failure. Persist the suspended state so a later approve()
-            // can continue. Both branches below settle the run's fate, so the
-            // finally-guard must not run either way.
-            $settled = true;
-
-            if ($handle !== null && $this->persister->suspend($handle, $approval->state)) {
-                return new AgentRunResult(
-                    outcome: AgentRunOutcome::AWAITING_APPROVAL,
-                    runUuid: $runUuid,
-                    steps: $trace->getSteps(),
-                    suspendedState: $approval->state,
-                );
-            }
-
-            // Fail-closed (ADR-092): an approval-gated tool is side-effecting.
-            // Without stored state — the store refused or errored, a concurrent
-            // cancel already terminated the row, or the run was never persisted
-            // at all (null handle) — there is nothing to resume, so promising
-            // an approval flow would strand the client. Applied on EVERY path
-            // since ADR-101 (the old code silently ignored a failed
-            // re-suspension AND announced awaiting-approval for unpersisted
-            // runs).
-            if ($handle !== null) {
-                // Ownership-guarded so a reclaimed queued run is not clobbered by
-                // this worker's fail-closed settle.
-                $this->persister->settleFailed($handle, $approval, $ownerGuard);
-            }
-            $this->logger?->error('Agent run could not be suspended for approval; no resume is possible', ['run' => $runUuid]);
-
-            return new AgentRunResult(
-                outcome: AgentRunOutcome::SUSPEND_FAILED,
-                runUuid: $runUuid,
-                steps: $trace->getSteps(),
-                error: $approval,
-            );
-        } catch (ToolInputRequiredException $input) {
-            // ADR-105: a called tool requires typed user input — control flow,
-            // not failure (the input sibling of the approval arm above). Must be
-            // caught before the guardrail pair and the generic Throwable so a
-            // suspension is never recorded as a failed run. Both branches settle
-            // the run's fate, so $settled=true keeps the finally-guard off — a
-            // successfully-suspended WAITING_FOR_INPUT row is non-terminal and
-            // must not be flipped to FAILED.
-            $settled = true;
-
-            if ($handle !== null && $this->persister->suspendForInput($handle, $input->state)) {
-                return new AgentRunResult(
-                    outcome: AgentRunOutcome::AWAITING_INPUT,
-                    runUuid: $runUuid,
-                    steps: $trace->getSteps(),
-                    suspendedState: $input->state,
-                );
-            }
-
-            // Fail-closed (ADR-092/105): without stored state — the store
-            // refused or errored, a concurrent cancel terminated the row, or the
-            // run was never persisted (null handle) — there is nothing to resume,
-            // so promising an input flow would strand the client.
-            if ($handle !== null) {
-                // Ownership-guarded so a reclaimed queued run is not clobbered by
-                // this worker's fail-closed settle.
-                $this->persister->settleFailed($handle, $input, $ownerGuard);
-            }
-            $this->logger?->error('Agent run could not be suspended for input; no resume is possible', ['run' => $runUuid]);
-
-            return new AgentRunResult(
-                outcome: AgentRunOutcome::SUSPEND_FAILED,
-                runUuid: $runUuid,
-                steps: $trace->getSteps(),
-                error: $input,
-            );
-        } catch (GuardrailViolationException|GuardrailApprovalRequiredException $guardrail) {
-            // ADR-085/086: a guardrail verdict is a policy outcome, not a
-            // failure — and an approval that was required but never obtained
-            // is not recorded as an outright denial (ADR-092).
-            if ($handle !== null) {
-                $settledOk = $this->persister->settlePolicyStopped(
-                    $handle,
-                    $guardrail,
-                    $guardrail instanceof GuardrailApprovalRequiredException
-                        ? AgentRunTerminationReason::APPROVAL_DENIED
-                        : AgentRunTerminationReason::POLICY_DENIED,
-                    $ownerGuard,
-                );
-                // Ownership-guarded on the queued path: a run reclaimed to
-                // another worker must not be flipped to a policy-stopped terminal
-                // by this zombie worker — stop as LEASE_LOST.
-                if ($ownerGuard !== null && !$settledOk) {
-                    $settled = true;
-                    $this->logger?->notice('Guardrail-stopped queued run was reclaimed; leaving it to its new owner', ['run' => $runUuid]);
-
-                    return new AgentRunResult(outcome: AgentRunOutcome::LEASE_LOST, runUuid: $runUuid, steps: $trace->getSteps(), error: $guardrail);
-                }
-            }
-            $settled = true;
-            $this->logger?->warning('Agent run blocked by guardrail', ['exception' => $guardrail, 'run' => $runUuid]);
-
-            return new AgentRunResult(
-                outcome: $guardrail instanceof GuardrailApprovalRequiredException
-                    ? AgentRunOutcome::GUARDRAIL_APPROVAL_REQUIRED
-                    : AgentRunOutcome::GUARDRAIL_BLOCKED,
-                runUuid: $runUuid,
-                steps: $trace->getSteps(),
-                guardrailClass: $guardrail->guardrail,
-                error: $guardrail,
-            );
-        } catch (RunLeaseLostException $leaseLost) {
-            // ADR-104: the reaper reclaimed this run (or a cancel/settle won) and
-            // it belongs to another worker now. Stop WITHOUT settling — any
-            // settle here could destroy the new owner's in-flight state. Not a
-            // failure of the run; a loss of ownership by this worker.
-            $settled = true;
-            $this->logger?->info('Agent run worker lost its lease; stopping without settling', ['run' => $runUuid]);
-
-            return new AgentRunResult(
-                outcome: AgentRunOutcome::LEASE_LOST,
-                runUuid: $runUuid,
-                steps: $trace->getSteps(),
-                error: $leaseLost,
-            );
-        } catch (Throwable $e) {
-            // A queued run's recover closure decides retry-vs-dead-letter and,
-            // when it handles the failure, settles the row itself and returns
-            // the replacement outcome. Only an interactive run (no recover) or a
-            // recover that declines (null) falls through to the default settle.
-            if ($recover !== null) {
-                $recovery = $recover($e, $trace->getSteps());
-                if ($recovery instanceof AgentRunResult) {
-                    $settled = true;
-
-                    return $recovery;
-                }
-            }
-
-            if ($handle !== null) {
-                $settledOk = $this->persister->settleFailed($handle, $e, $ownerGuard);
-                if ($ownerGuard !== null && !$settledOk) {
-                    $settled = true;
-                    $this->logger?->notice('Failed queued run was reclaimed; leaving it to its new owner', ['run' => $runUuid]);
-
-                    return new AgentRunResult(outcome: AgentRunOutcome::LEASE_LOST, runUuid: $runUuid, steps: $trace->getSteps(), error: $e);
-                }
-            }
-            $settled = true;
-            $this->logger?->error('Agent run failed', ['exception' => $e, 'run' => $runUuid]);
-
-            return new AgentRunResult(
-                outcome: AgentRunOutcome::FAILED,
-                runUuid: $runUuid,
-                steps: $trace->getSteps(),
-                error: $e,
-            );
-        } finally {
-            // A live-stream observer dying mid-run (client disconnect) can
-            // abandon the run before any branch settles it; mark it failed so
-            // no run is left stuck RUNNING. Mirrors StreamingDispatcher's
-            // finally-block settle. Guarded by $settled: a suspended run is
-            // WAITING_FOR_APPROVAL or WAITING_FOR_INPUT — both non-terminal — and
-            // settling it here would destroy its resumable state.
-            if ($handle !== null && !$settled) {
-                // Ownership-guarded so this safety-net settle cannot clobber a
-                // queued run the reaper reclaimed to another worker.
-                $this->persister->settleFailed($handle, new RuntimeException('Agent run did not complete'), $ownerGuard);
-            }
-        }
-    }
 }
